Add files via upload
A program to calculate the sum of the numbers entered by a user. The program repeatedly requests input values. The user wants to enter one number at a time and add 0 as the end value to indicate the end. This program then displays the sum of all the values entered.