len function - python
A program to show the longest string entered by user.this program ask for 2 Strings as inputs and use the longest str function you defined, to determine the longest string and then output it. using the len() function to get the length of a string.A program to show the longest string entered by user.this program ask for 2 Strings as inputs and use the longest str function you defined, to determine the longest string and then output it. using the len() function to get the length of a string.